LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/modal-dialog-selection-expected.txt

123456789A

FAIL By default, text inside a modal dialog can be selected assert_equals: expected "345678" but got ""
PASS 'user-select: none' prevents text from being selected
